+++ catalogue/rdbms/RdbmsTapePoolCatalogue.hpp
@@ -18,6 +18,7 @@
 #pragma once
 
 #include <list>
+#include <set>
 #include <memory>
 #include <optional>
 #include <string>
@@ -119,8 +120,8 @@ private:
   friend class RdbmsTapeCatalogue;
   std::optional<uint64_t> getTapePoolId(rdbms::Conn &conn, const std::string &name) const;
 
-  std::optional<std::string> getTapePoolSupplySources(rdbms::Conn &conn, const std::string &tapePoolName) const;
-  std::optional<std::string> getTapePoolSupplyDestinations(rdbms::Conn &conn, const std::string &tapePoolName) const;
+  std::set<std::string> getTapePoolSupplySources(rdbms::Conn &conn, const std::string &tapePoolName) const;
+  std::set<std::string> getTapePoolSupplyDestinations(rdbms::Conn &conn, const std::string &tapePoolName) const;
 
   std::vector<std::string> verifyTapePoolSupply(rdbms::Conn &conn, const std::string &supply);
   void populateSupplyTable(rdbms::Conn &conn, std::string tapePoolName, std::vector<std::string> verified_matches);